Integrate database to store and manage property listings
Database has been integrated for agents to be able to add, update and delete property listings on to the property database. 
Agents can now view the property listings with access to the database on the property window.
Agents also have better and improved usability features, for example the prevention of the same property being added more than once, or the same property being deleted more than once. This is thanks to the unique property ID generated for each property listing to help prevent the case.
Agents are required to enter all information required on the addProperty, updateProperty or deleteProperty windows to successfully submit the request, otherwise request is void.